requirements.txt
flask==3.0.0
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas>=2.2.0
openai==1.3.7
python-dotenv==1.0.0
